requirement.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.6
# supabase-py>=1.0.0

SQLAlchemy>=1.4.0
# passlib[bcrypt]>=1.7.4
passlib[argon2]
python-jose[cryptography]>=3.3.0

#lib for schemas
pydantic[email]

pytest 
pytest-mock
httpx>=0.27.0
kafka-python>=2.0.2
praw>=7.7.0